requirements.txt
aiosqlite==0.17.0
fastapi==0.65.2
httpx==0.16.1
python-dotenv==0.17.0
reasoner-pydantic==1.2.0.3
uvicorn==0.13.4
